experiment/models/requirements.txt
python-doctr[viz,html,contrib]
torch
torchvision
tqdm
Pillow
numpy